Fix the GestureTest example
Resolve the interrupt number with digitalPinToInterrupt() instead of
hardcoding it to "0". This is the recommended method from
https://www.arduino.cc/en/Reference/attachInterrupt

It should work with post-2014 releases of any Arduino IDE version.
(1.0.x, 1.5.x, 1.6.x)

Also declare the isr_flag as volatile. It will not make a difference
here but can break code if this example is extended further.